specify liquibase integration when running through ant (DAT-18608) (#6386)
specify liquibase integration when running through ant
