+++ liquibase-standard/src/test/resources/liquibase/integration/ant/test/antlib.xml
@@ -4,6 +4,8 @@
     <typedef name="columnExists" classname="liquibase.integration.ant.test.ColumnExistsCondition"/>
     <typedef name="rowCountEquals" classname="liquibase.integration.ant.test.RowCountEqualsCondition"/>
     <typedef name="tagExists" classname="liquibase.integration.ant.test.LiquibaseTagExistsCondition"/>
+    <typedef name="analyticsWebserverPostBodyIsAnt" classname="liquibase.integration.ant.test.AnalyticsWebserverPostBodyIsAnt"/>
+    <typedef name="antTestAnalyticsWebserver" classname="liquibase.integration.ant.test.AntTestAnalyticsWebserver"/>
 
     <macrodef name="assertTableExists" backtrace="false">
         <attribute name="driver"/>
@@ -103,6 +105,27 @@
             <au:assertFalse message="@{message}">
                 <current:tagExists driver="@{driver}" url="@{url}" user="@{user}" password="@{password}" tag="@{tag}"/>
             </au:assertFalse>
+        </sequential>
+    </macrodef>
+
+    <macrodef name="startTestAnalyticsWebserver" backtrace="false">
+        <sequential>
+            <current:antTestAnalyticsWebserver up="true"/>
+        </sequential>
+    </macrodef>
+
+    <macrodef name="stopTestAnalyticsWebserver" backtrace="false">
+        <sequential>
+            <current:antTestAnalyticsWebserver up="false"/>
+        </sequential>
+    </macrodef>
+
+    <macrodef name="assertAnalyticsPostBodyIsAnt" backtrace="false">
+        <attribute name="message" default="Assertion failed"/>
+        <sequential>
+            <au:assertTrue message="@{message}">
+                <current:analyticsWebserverPostBodyIsAnt/>
+            </au:assertTrue>
         </sequential>
     </macrodef>
 </antlib>
